+++ cmd/hidden_lake/service/pkg/app/errors.go
@@ -25,4 +25,5 @@ var (
 	ErrInitConfig        = &SAppError{"init config"}
 	ErrSetParallelNull   = &SAppError{"set parallel = 0"}
 	ErrGetParallel       = &SAppError{"get parallel"}
+	ErrCreateAnonNode    = &SAppError{"create anon node"}
 )