+++ pkg/wrapper/doc.go
@@ -1,2 +1,0 @@
-// Package wrapper is used to create objects with the possibility of their subsequent rewriting.
-package wrapper