+++ pkg/wrapper/types.go
@@ -1,6 +1,0 @@
-package wrapper
-
-type IWrapper interface {
-	Get() interface{}
-	Set(interface{}) IWrapper
-}