+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Update `cmd/hidden_lake/applications/messenger`: delete share_enabled option, requestID
 - Update `examples/anon_messenger/group`: change group chat to use one private key
 - Update `go-peer`: all packages with custom errors now have errors.go file
+- Update `pkg`: delete wrapper package
 
 ### BUG FIXES